requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
Django==4.0.2
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
gunicorn==20.1.0
Pillow==9.0.1
soupsieve==2.3.1
sqlparse==0.4.2
